Refactor AwsUrlUtils
